docs(sitemap) create sitemap.xml during build
Also splits out `nav` and `sidebar` sections of the config into their own files.
